made lots of small improvements
